# contracts/agents-api/core.openapi.yaml
basePath: /
definitions:
...
  api.CoreAPIError:
    properties:
      code:
        type: string
        x-nullable: true
      details:
        description: |-
          Details contains only documented, Core-owned facts: string, finite number,
          boolean, null or string array values. Never include request echoes, secrets
          or native/provider error text. Empty or invalid details are omitted.
        type: object
      message:
        type: string
      param:
        type: string
        x-nullable: true
      type:
        type: string
    required:
      - message
      - type
    type: object
  api.CoreErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.CoreAPIError'
    required:
      - error
    type: object
...
  api.SandboxAllocationList:
    properties:
      data:
        items:
          $ref: '#/definitions/store.RuntimeNodeAllocation'
        type: array
    type: object
  api.SandboxDeploymentChangeInput:
    properties:
      e2b:
        $ref: '#/definitions/api.SandboxE2BInput'
      expected_generation:
        type: integer
      provider:
        type: string
      resources:
        allOf:
          - $ref: '#/definitions/sandbox.Resources'
        description: |-
          Per-sandbox limits, required for Docker and microsandbox. E2B may omit
          them; Core then uses the validated template build's cpus and memory_mib.
      runtime:
        $ref: '#/definitions/sandbox.RuntimeRelease'
    required:
      - expected_generation
    type: object
  api.SandboxDeploymentInput:
    properties:
      e2b:
        $ref: '#/definitions/api.SandboxE2BInput'
      expected_generation:
        type: integer
      provider:
        type: string
      resources:
        allOf:
          - $ref: '#/definitions/sandbox.Resources'
        description: |-
          Per-sandbox limits, required for Docker and microsandbox. E2B may omit
          them; Core then uses the validated template build's cpus and memory_mib.
      runtime:
        $ref: '#/definitions/sandbox.RuntimeRelease'
    required:
      - expected_generation
    type: object
  api.SandboxE2BInput:
    properties:
      api_key:
        type: string
      template:
        type: string
    type: object
  api.SandboxEnrollmentToken:
    properties:
      enrollment_id:
        description: Public, non-secret handle of this command; never a credential. The node it registers reports the same value as enrollment_id.
        type: string
      expires_at:
        type: string
      token:
        type: string
    type: object
  api.SandboxEnrollmentTokenRequest:
    properties:
      max_active:
        type: integer
      max_retained:
        description: Docker never suspends, so Core replaces this with max_active; microsandbox uses both limits.
        type: integer
    type: object
  api.SandboxMutationResponse:
    properties:
      deleted:
        type: boolean
      id:
        type: string
      updated:
        type: boolean
    type: object
  api.SandboxNodeList:
    properties:
      data:
        items:
          $ref: '#/definitions/store.RuntimeNode'
        type: array
    type: object
...
  sandbox.DeploymentSpec:
    properties:
      resources:
        $ref: '#/definitions/sandbox.Resources'
      runtime:
        $ref: '#/definitions/sandbox.RuntimeRelease'
    type: object
  sandbox.Resources:
    properties:
      cpus:
        type: integer
      environment_disk_mib:
        type: integer
      memory_mib:
        type: integer
      root_disk_mib:
        type: integer
    type: object
  sandbox.RuntimeRelease:
    properties:
      firmware_sha256:
        type: string
      image_id:
        type: string
      image_manifest_digest:
        type: string
      microsandbox_ref:
        type: string
      runtime_sha256:
        type: string
      source_commit:
        type: string
    type: object
...
  store.NodeHostHistory:
    properties:
      points:
        items:
          $ref: '#/definitions/store.NodeHostHistoryPoint'
        type: array
      resolution_seconds:
        type: integer
    type: object
...
  store.RuntimeDeploymentView:
    properties:
      core_url:
        description: 'Read-only: the installation public URL (OAC_PUBLIC_URL), which nodes and sandboxes use to reach Core. The deployment API does not accept it.'
        type: string
      e2b:
        $ref: '#/definitions/store.SandboxE2BView'
      generation:
        type: integer
      installation_id:
        type: string
      mode:
        type: string
      owner_epoch:
        type: integer
      provider:
        type: string
      reset:
        allOf:
          - $ref: '#/definitions/store.SandboxResetView'
        x-nullable: true
      resources:
        $ref: '#/definitions/store.SandboxDeploymentResources'
      rollout:
        $ref: '#/definitions/store.SandboxRollout'
      specification:
        $ref: '#/definitions/sandbox.DeploymentSpec'
      specification_digest:
        type: string
      suspension:
        allOf:
          - $ref: '#/definitions/store.SandboxSuspensionView'
        description: Idle suspension policy; microsandbox only, otherwise null.
        x-nullable: true
    type: object
  store.RuntimeNode:
    properties:
      active:
        type: integer
      available_disk_bytes:
        type: integer
      available_memory_bytes:
        type: integer
      cleanup_pending:
        type: integer
      core_url:
        description: |-
          The Core address this node enrolled with. A node whose address differs
          from the installation public URL receives no new sandboxes; re-add it.
        type: string
      cpu_count:
        type: integer
      created_at:
        type: string
      diagnostic:
        description: Fixed reason for the last reported unreadiness; absent while the provider is ready. Clients treat an unknown value as provider_unavailable.
        enum:
          - provider_unavailable
          - docker_unavailable
          - docker_limits_unsupported
          - runtime_download_failed
          - runtime_image_unavailable
          - kvm_unavailable
          - microsandbox_artifacts_unavailable
          - capacity_insufficient
        type: string
      enrollment_id:
        description: The enrollment_id of the command that registered this node (POST /core/v1/sandbox/enrollment-tokens); null for nodes enrolled before Core recorded it.
        type: string
        x-nullable: true
      id:
        type: string
      last_seen_at:
        type: string
      max_active:
        type: integer
      max_retained:
        type: integer
      name:
        type: string
      online:
        type: boolean
      provider:
        type: string
      provider_ready:
        type: boolean
      reserved:
        type: integer
      retained:
        type: integer
      rollout:
        $ref: '#/definitions/store.SandboxNodeRollout'
      running:
        type: integer
      snapshots:
        type: integer
    type: object
...
  store.RuntimeNodeDetail:
    properties:
      active:
        type: integer
      available_disk_bytes:
        type: integer
      available_memory_bytes:
        type: integer
      cleanup_pending:
        type: integer
      core_url:
        description: |-
          The Core address this node enrolled with. A node whose address differs
          from the installation public URL receives no new sandboxes; re-add it.
        type: string
      cpu_count:
        type: integer
      created_at:
        type: string
      diagnostic:
        description: Fixed reason for the last reported unreadiness; absent while the provider is ready. Clients treat an unknown value as provider_unavailable.
        enum:
          - provider_unavailable
          - docker_unavailable
          - docker_limits_unsupported
          - runtime_download_failed
          - runtime_image_unavailable
          - kvm_unavailable
          - microsandbox_artifacts_unavailable
          - capacity_insufficient
        type: string
      enrollment_id:
        description: The enrollment_id of the command that registered this node (POST /core/v1/sandbox/enrollment-tokens); null for nodes enrolled before Core recorded it.
        type: string
        x-nullable: true
      history:
        $ref: '#/definitions/store.NodeHostHistory'
      host:
        $ref: '#/definitions/store.RuntimeNodeHost'
      id:
        type: string
      last_seen_at:
        type: string
      max_active:
        type: integer
      max_retained:
        type: integer
      name:
        type: string
      online:
        type: boolean
      provider:
        type: string
      provider_ready:
        type: boolean
      reserved:
        type: integer
      retained:
        type: integer
      rollout:
        $ref: '#/definitions/store.SandboxNodeRollout'
      running:
        type: integer
      snapshots:
        type: integer
    type: object
  store.RuntimeNodeHost:
    properties:
      available_disk_bytes:
        type: integer
        x-nullable: true
      available_memory_bytes:
        type: integer
        x-nullable: true
      cpu_utilization:
        type: number
        x-nullable: true
      effective_cpu_cores:
        type: number
        x-nullable: true
      observed_at:
        type: string
        x-nullable: true
      total_memory_bytes:
        type: integer
        x-nullable: true
    type: object
  store.RuntimeNodeUpdate:
    properties:
      max_active:
        type: integer
      max_retained:
        description: Docker never suspends, so Core replaces this with max_active; microsandbox uses both limits.
        type: integer
      name:
        type: string
    type: object
  store.SandboxDeploymentResources:
    properties:
      allocations:
        type: integer
      pending:
        type: integer
    type: object
  store.SandboxE2BTemplateBuildView:
    properties:
      resources:
        $ref: '#/definitions/store.SandboxTemplateResources'
      status:
        description: Build status at selection time; validation admits only ready builds.
        type: string
        x-nullable: true
    type: object
  store.SandboxE2BView:
    properties:
      credential_configured:
        type: boolean
      template:
        type: string
      template_build:
        allOf:
          - $ref: '#/definitions/store.SandboxE2BTemplateBuildView'
        description: The fixed template build as Core read it when this selection was saved.
    type: object
...
  store.SandboxResetView:
    properties:
      clear:
        type: string
      deadline_at:
        type: string
        x-nullable: true
      forced_at:
        type: string
        x-nullable: true
      remaining:
        $ref: '#/definitions/store.SandboxResetRemaining'
      requested_at:
        type: string
    type: object
  store.SandboxRollout:
    properties:
      nodes:
        allOf:
          - $ref: '#/definitions/store.SandboxRolloutNodes'
        x-nullable: true
      previous_generation_sandboxes:
        type: integer
      state:
        enum:
          - settled
          - preparing
        type: string
    type: object
...
  store.SandboxSuspensionView:
    properties:
      idle_seconds:
        type: integer
      retention_seconds:
        type: integer
    type: object
...
info:
  contact: {}
  description: Deployment and operations routes under /core/v1 for Core Web's server and operator scripts. Every operation requires the Core key; Project API keys and machine credentials are not accepted.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenAgentCore Core API
  version: "1"
paths:
...
  /core/v1/sandbox/deployment:
    get:
      description: Core key only. Does not grant project resource access. Responses contain only explicit safe fields. E2B template_build values are those Core read when the selection was saved; this read does not call E2B.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.RuntimeDeploymentView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: Retrieve sandbox deployment
      tags:
        - Sandbox Manager
    post:
      consumes:
        - application/json
      description: Selects a provider, enforced resource limits and pinned Runtime release. Core derives the deployment's core_url from the installation public URL and rejects a core_url member with 400. E2B returns 409 sandbox_configuration_error while the public URL is loopback. E2B credentials are write-only. E2B may omit resources to adopt the validated template build's CPU and memory, returned in specification.resources. Requires explicit expected_generation, including zero at first setup. Stale retries reject before provider validation. An identical selection at the current generation is a no-op; differing selections and file-managed deployments reject. This does not create compute or execute work.
      parameters:
        - description: Deployment selection
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/api.SandboxDeploymentInput'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.RuntimeDeploymentView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: Initialize the deployment sandbox provider
      tags:
        - Sandbox Manager
    put:
      consumes:
        - application/json
      description: 'Requires the observed generation, the same backend type and no active reset. E2B same-team changes apply online: allocations retain immutable generation and current credentials; omitted api_key preserves it, explicit submission including the same key verifies and advances generation. Other teams require explicit reset. Node providers retain the zero-resource guard and retire old nodes/tokens on change. Core rejects core_url input. Never automatically replay an uncertain write; rollout.state is the authoritative preparation polling signal.'
      parameters:
        - description: Replacement deployment selection
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/api.SandboxDeploymentChangeInput'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.RuntimeDeploymentView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: Change the sandbox deployment configuration
      tags:
        - Sandbox Manager
...
  /core/v1/sandbox/enrollment-tokens:
    post:
      consumes:
        - application/json
      description: Core key only. Does not grant project resource access. Responses contain only explicit safe fields.
      parameters:
        - description: Request
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/api.SandboxEnrollmentTokenRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.SandboxEnrollmentToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: Create a ten-minute one-use node enrollment token
      tags:
        - Sandbox Manager
  /core/v1/sandbox/nodes:
    get:
      description: Core key only. Does not grant project resource access. Responses contain only explicit safe fields.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SandboxNodeList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: List deployment sandbox nodes
      tags:
        - Sandbox Manager
  /core/v1/sandbox/nodes/{node_id}:
    delete:
      description: Core key only. Does not grant project resource access. Responses contain only explicit safe fields.
      parameters:
        - description: Sandbox node UUID
          in: path
          name: node_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SandboxMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: Remove a sandbox node with no retained resources
      tags:
        - Sandbox Manager
    get:
      description: Core key only. Complete UTC buckets. Missing host measurements and offline history are null; reads never sample or backfill.
      parameters:
        - description: Sandbox node UUID
          in: path
          name: node_id
          required: true
          type: string
        - description: Time range (default 1h)
          enum:
            - 1h
            - 6h
            - 24h
          in: query
          name: range
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.RuntimeNodeDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: Retrieve sandbox node and host history
      tags:
        - Sandbox Manager
    patch:
      consumes:
        - application/json
      description: Core key only. Does not grant project resource access. Responses contain only explicit safe fields.
      parameters:
        - description: Sandbox node UUID
          in: path
          name: node_id
          required: true
          type: string
        - description: Request
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/store.RuntimeNodeUpdate'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SandboxMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: Update sandbox node name and capacity
      tags:
        - Sandbox Manager
  /core/v1/sandbox/nodes/{node_id}/allocations:
    get:
      description: Core key only. Does not grant project resource access. Responses contain only explicit safe fields.
      parameters:
        - description: Sandbox node UUID
          in: path
          name: node_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SandboxAllocationList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.CoreErrorResponse'
      security:
        - DeploymentAdminAuth: []
      summary: List retained allocations on a sandbox node
      tags:
        - Sandbox Manager
...
schemes:
  - http
  - https
securityDefinitions:
  DeploymentAdminAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"